apps/api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlmodel==0.0.22
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.43.0
httpx==0.27.0
